refactor: split functional.py into api/ sub-package
- Create src/agedi/api/_registry.py with model registry and builder helpers
- Create src/agedi/api/_display.py with Rich display utilities
- Create src/agedi/api/diffusion.py with create_diffusion and load_diffusion
- Create src/agedi/api/dataset.py with create_dataset
- Create src/agedi/api/training.py with create_trainer, train, train_from_atoms, train_from_config
- Create src/agedi/api/sampling.py with sample
- Create src/agedi/api/prediction.py with predict
- Create src/agedi/api/__init__.py re-exporting all public names
- Reduce functional.py to a backward-compat shim

Agent-Logs-Url: https://github.com/nronne/agedi/sessions/ef552323-6338-4cc8-9074-42f6908cd465
