+++ src/agedi/api/__init__.py
@@ -1,0 +1,29 @@
+"""Public API for AGeDi.
+
+Re-exports all public symbols from the api sub-modules so that
+``from agedi.api import X`` works for every user-facing name.
+"""
+
+from ._registry import register_model as register_model
+from .dataset import create_dataset as create_dataset
+from .diffusion import create_diffusion as create_diffusion
+from .diffusion import load_diffusion as load_diffusion
+from .prediction import predict as predict
+from .sampling import sample as sample
+from .training import create_trainer as create_trainer
+from .training import train as train
+from .training import train_from_atoms as train_from_atoms
+from .training import train_from_config as train_from_config
+
+__all__ = [
+    "create_dataset",
+    "create_diffusion",
+    "create_trainer",
+    "load_diffusion",
+    "predict",
+    "register_model",
+    "sample",
+    "train",
+    "train_from_atoms",
+    "train_from_config",
+]
